+++ src/GRA/ClaimType.cs
@@ -1,8 +1,8 @@
-﻿
-namespace GRA
+﻿namespace GRA
 {
     public static class ClaimType
     {
+        public static readonly string AuthenticatedAt = "GRA.AuthenticatedAt";
         public static readonly string BranchId = "GRA.BranchId";
         public static readonly string Permission = "GRA.Permission";
         public static readonly string SiteId = "GRA.SiteId";