+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - New mail icon in navigation is red and solid if there are new messages, grey and wireframe otherwise
 - Upgrade ImageSharp from v1.0.0-beta0003 to v1.0.0-beta0005
 - Move site.css for customizing site styles to shared folder
+- Only redirect to the sign in page with the session expired message if the participant has authenticated within the last two hours
 
 ### Fixed
 - Fix #377 Adding a challenge task causes an error